+++ packages/chain-adapters/src/ChainAdapterManager.test.ts
@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { ChainTypes } from '@shapeshiftoss/types'
 
-import { UnchainedUrls } from './ChainAdapterManager'
-import { ChainAdapterManager } from './ChainAdapterManager'
+import { ChainAdapterManager, UnchainedUrls } from './ChainAdapterManager'
 import * as ethereum from './ethereum'
 
 const getCAM = (opts?: UnchainedUrls) => {
@@ -112,13 +111,12 @@ describe('ChainAdapterManager', () => {
 
   describe('getSupportedAdapters', () => {
     it('should return array of adapter classes', () => {
-      // @ts-ignore
       expect(getCAM().getSupportedAdapters()).toStrictEqual([expect.any(Function)])
     })
   })
 
   describe('byChainId', () => {
-    it('should find a supported chain adapter', async () => {
+    it('should find a supported chain adapter', () => {
       const cam = new ChainAdapterManager({})
       // @ts-ignore
       cam.addChain(ChainTypes.Bitcoin, () => ({
@@ -129,17 +127,17 @@ describe('ChainAdapterManager', () => {
         getChainId: () => 'eip155:1'
       }))
 
-      await expect(cam.byChainId('eip155:1')).resolves.toBeTruthy()
+      expect(cam.byChainId('eip155:1')).toBeTruthy()
     })
 
-    it('should throw an error for an invalid ChainId', async () => {
+    it('should throw an error for an invalid ChainId', () => {
       const cam = new ChainAdapterManager({})
-      await expect(cam.byChainId('fake:chainId')).rejects.toThrow('invalid')
+      expect(() => cam.byChainId('fake:chainId')).toThrow('Chain [fake:chainId] is not supported')
     })
 
-    it('should throw an error if there is no supported adapter', async () => {
+    it('should throw an error if there is no supported adapter', () => {
       const cam = new ChainAdapterManager({})
-      await expect(cam.byChainId('eip155:1')).rejects.toThrow('not supported')
+      expect(() => cam.byChainId('eip155:1')).toThrow('Chain [eip155:1] is not supported')
     })
   })
 })