+++ packages/swapper/src/swappers/zrx/ZrxApprovalNeeded/ZrxApprovalNeeded.ts
@@ -1,3 +1,4 @@
+import { fromAssetId, getFeeAssetIdFromAssetId } from '@shapeshiftoss/caip'
 import { ApprovalNeededOutput, SupportedChainIds } from '@shapeshiftoss/types'
 
 import { ApprovalNeededInput, SwapError, SwapErrorTypes } from '../../../api'
@@ -13,6 +14,8 @@ export async function ZrxApprovalNeeded(
 ): Promise<ApprovalNeededOutput> {
   const { sellAsset } = quote
 
+  const { assetReference: sellAssetErc20Address } = fromAssetId(sellAsset.assetId)
+
   try {
     if (sellAsset.chainId !== 'eip155:1') {
       throw new SwapError('[ZrxApprovalNeeded] - sellAsset chainId is not supported', {
@@ -21,8 +24,8 @@ export async function ZrxApprovalNeeded(
       })
     }
 
-    // No approval needed for trading eth itself
-    if (sellAsset.assetId === 'eip155:1/slip44:60') {
+    // No approval needed for selling a fee asset
+    if (sellAsset.assetId === getFeeAssetIdFromAssetId(sellAsset.assetId)) {
       return { approvalNeeded: false }
     }
 
@@ -32,8 +35,8 @@ export async function ZrxApprovalNeeded(
     const bip44Params = adapter.buildBIP44Params({ accountNumber })
     const receiveAddress = await adapter.getAddress({ wallet, bip44Params })
 
-    if (!quote.sellAsset.tokenId || !quote.allowanceContract) {
-      throw new SwapError('[ZrxApprovalNeeded] - tokenId and allowanceTarget are required', {
+    if (!quote.allowanceContract) {
+      throw new SwapError('[ZrxApprovalNeeded] - allowanceTarget is required', {
         code: SwapErrorTypes.VALIDATION_FAILED,
         details: { chainId: sellAsset.chainId }
       })
@@ -42,7 +45,7 @@ export async function ZrxApprovalNeeded(
     const allowanceResult = await getERC20Allowance({
       web3,
       erc20AllowanceAbi,
-      tokenId: quote.sellAsset.tokenId,
+      sellAssetErc20Address,
       spenderAddress: quote.allowanceContract,
       ownerAddress: receiveAddress
     })