+++ packages/swapper/src/swappers/zrx/ZrxSwapper.ts
@@ -19,11 +19,12 @@ import {
   ExecuteTradeInput,
   GetTradeQuoteInput,
   Swapper,
-  Trade,
-  TradeQuote
+  TradeQuote,
+  ZrxTrade
 } from '../../api'
 import { getZrxMinMax } from './getZrxMinMax/getZrxMinMax'
 import { getZrxTradeQuote } from './getZrxTradeQuote/getZrxTradeQuote'
+import { UNSUPPORTED_ASSETS } from './utils/blacklist'
 import { getUsdRate } from './utils/helpers/helpers'
 import { ZrxApprovalNeeded } from './ZrxApprovalNeeded/ZrxApprovalNeeded'
 import { ZrxApproveInfinite } from './ZrxApproveInfinite/ZrxApproveInfinite'
@@ -47,7 +48,7 @@ export class ZrxSwapper implements Swapper {
     return SwapperType.Zrx
   }
 
-  async buildTrade(args: BuildTradeInput): Promise<Trade<SupportedChainIds>> {
+  async buildTrade(args: BuildTradeInput): Promise<ZrxTrade<SupportedChainIds>> {
     return zrxBuildTrade(this.deps, args)
   }
 
@@ -55,7 +56,7 @@ export class ZrxSwapper implements Swapper {
     return getZrxTradeQuote(input)
   }
 
-  async getUsdRate(input: Pick<Asset, 'symbol' | 'tokenId'>): Promise<string> {
+  async getUsdRate(input: Pick<Asset, 'symbol' | 'assetId'>): Promise<string> {
     return getUsdRate(input)
   }
 
@@ -79,10 +80,15 @@ export class ZrxSwapper implements Swapper {
 
   filterBuyAssetsBySellAssetId(args: BuyAssetBySellIdInput): AssetId[] {
     const { assetIds = [], sellAssetId } = args
-    return assetIds.filter((id) => id.startsWith('eip155:1') && sellAssetId?.startsWith('eip155:1'))
+    return assetIds.filter(
+      (id) =>
+        id.startsWith('eip155:1') &&
+        sellAssetId?.startsWith('eip155:1') &&
+        !UNSUPPORTED_ASSETS.includes(id)
+    )
   }
 
   filterAssetIdsBySellable(assetIds: AssetId[] = []): AssetId[] {
-    return assetIds.filter((id) => id.startsWith('eip155:1'))
+    return assetIds.filter((id) => id.startsWith('eip155:1') && !UNSUPPORTED_ASSETS.includes(id))
   }
 }